cookie_factory

Macro gen_at_rel_offset

macro_rules! gen_at_rel_offset {
    (($i:expr, $idx:expr), $rel_offset:expr, $f:ident( $($args:tt)* )) => { ... };
    (($i:expr, $idx:expr), $rel_offset:expr, $submac:ident!( $($args:tt)* )) => { ... };
}

gen_at_offset!(usize, I -> Result<I,E>) => I -> Result<I,E> Combinator to call generator at a relative offset.

Wraps another generator and calls it using a different index from the current position. If this combinator succeeds, it returns the current index (instead of the one returned by the child generator). If the child generator fails, returns the error.