cookie_factory

Trait BackToTheBuffer

pub trait BackToTheBuffer: Write {
    // Required method
    fn reserve_write_use<Tmp, Gen: Fn(WriteContext<Self>) -> Result<(WriteContext<Self>, Tmp), GenError>, Before: Fn(WriteContext<Self>, Tmp) -> GenResult<Self>>(
        s: WriteContext<Self>,
        reserved: usize,
        gen: &Gen,
        before: &Before,
    ) -> Result<WriteContext<Self>, GenError>
       where Self: Sized;
}

Trait for Write types that allow skipping and reserving a slice, then writing some data, then write something in the slice we reserved using the return for our data write.

Required Methods

fn reserve_write_use<Tmp, Gen: Fn(WriteContext<Self>) -> Result<(WriteContext<Self>, Tmp), GenError>, Before: Fn(WriteContext<Self>, Tmp) -> GenResult<Self>>( s: WriteContext<Self>, reserved: usize, gen: &Gen, before: &Before, ) -> Result<WriteContext<Self>, GenError>
where Self: Sized,