cookie_factory

Trait SerializeFn

pub trait SerializeFn<W>: Fn(WriteContext<W>) -> GenResult<W> { }

Trait for serializing functions

Serializing functions take one input W that is the target of writing and return an instance of cookie_factory::GenResult.

This trait is implemented for all Fn(W) -> GenResult<W>.